requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
idna==3.11
markdownify==0.14.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
python-dotenv==1.0.1
